Split trusted stdlib type refinements
