+++ crates/masm-decompiler/src/types/mod.rs
@@ -6,6 +6,7 @@ mod declared;
 mod domain;
 mod inter;
 mod intra;
+mod stdlib;
 mod summary;
 
 pub(crate) use declared::{declared_summary_for_proc, declared_summary_for_proc_with_arity};
